Change the way that variables are managed for opsview_contact resources.
Previously, the only variable that could be managed via opsview_contact
was 'EMAIL' via the 'email' property, and since opsview notification
variables can be user-defined this was not sufficient. This change
replaces the 'email' property with 'variables' which accepts a hash
containing variable 'name' => 'value' pair. Now, opsview_contact can
support *any* variables.

Also, Opsview will automatically add all variables to all contacts,
and we don't want to be forced to manage every variable in our
manifests in order to avoid puppet re-applying changes on every run.
So, we only check for insync? on variables that we are actually
managing in the manifest.