Set more default values in types
Specifically set array properties by default to [] (the empty array) and
notification_interval to 0. Don't set a default for hosttemplate's
managementurls, though; that doesn't currently work, because the
managementurls object should be a list of hashes with "name" and "url"
keys, but that doesn't seem to be taken into account by the
hosttemplate provider.